refactor: split ssh_bootstrap tests to clear max-lines ratchet
ssh_bootstrap.rs was 1304 lines (baseline 1145). Move unit tests to
ssh_bootstrap_tests.rs via #[path] like other alera-cli modules.
